Delete code to transform xlsx to json
