--- constants.py
@@ -1,2 +1,0 @@
-DATASETS_LIST_PATH = "/home/cangonin/Documents/datasets_list/bioacoustics_datasets.xlsx"
-OUTPUT_FOLDER = "/home/cangonin/Documents/datasets_list/croissant"  # Folder where the json files will be written
